STYLE: 코드 스타일 전체 정리
