+++ data/search_place.py
@@ -4,14 +4,14 @@ import requests
 from assets.api_len import api_length_dict
 
 
-def search_restaurant(place: str, city: str):
+def search_restaurant(place: str, city: str):   
     i = 1
-    data = get_restaurants_info(city, i)
-    for res in data:
-        print(res['BIZPLC_NM'])
-    # while i <= api_length_dict[city]:
-    #     data = get_restaurants_info(city, i)
-    #     for res in data:
+    data = []
+    
+    while i <= api_length_dict[city]:
+        data.append(get_restaurants_info(city, i))
+        i += 1
+    return data
 
 def get_restaurants_info(city: str, page_index: int) -> List:
     BASE_URL = "https://openapi.gg.go.kr/GENRESTRT?Type=json"
@@ -23,6 +23,7 @@ def get_restaurants_info(city: str, page_index: int) -> List:
 
 if __name__ == "__main__":
     data = search_restaurant("잉꼬칼국수", "구리시")
+    print(data)
     # import pprint
 
     # if data:
